Version: 2.7.D
- Small fixes, ensuring proper casting, clearing warnings, comments, in general just housekeeping.

+++ src/Assets/Asset_Image.cpp
@@ -75,7 +75,7 @@ void Asset_Image::fill(const glm::uvec4 primaryColor, const glm::uvec4 secondary
 				{ GLubyte(secondaryColor.x), GLubyte(secondaryColor.y), GLubyte(secondaryColor.z), GLubyte(secondaryColor.w) }
 			};
 			bool cFlip = false;
-			for (size_t pixel = 0, depth = -1; pixel < pixelCount; ++pixel) {
+			for (size_t pixel = 0, depth = size_t(-1); pixel < pixelCount; ++pixel) {
 				// Flip color if we've drawn enough of the checker
 				if ((pixel % checkerSize) == 0)
 					cFlip = !cFlip;
